Reverts previous HMM_FastInverseSquareRoot revert :P
Hlkpf3 on Mārtiņš Možeiko's comment:

i don't really want to be that guy here.. but it's the internet
and i have no free will: the new built-ins are using 'the function',
so they are not faster than 'the function' - they're just quicker at it.
also, many instruction sets don't offer that functionality yet.
and for some it doesn't make any sense. playstations and the nintendo wii,
for instance, were using powerpc - no. the nintendo wii u also doesn't
have it. the "new 3ds" should in theory be the first nintendo-device with
this intrinsic thanks to ARM11 NEON. can someone confirm that? heck,
some MPUs still don't even have floating-point units. i have heard of
a case where the energy-consumption of a sensor-network for
indoor-localisation could be halved by replacing the sqrt they used
for Pythagoras with this neat buddy here.﻿